Src - Refactor Snow-17 csv to a pandas dataframe
Deprecating the CSV handler and refactor to a pandas dataframe helper
class. This is the result of fetching the data now straight from the
DB.
